% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{NGCHMDemoData}
\alias{NGCHMDemoData}
\title{NGCHMDemoData}
\description{
This package provides several relatively large datasets that can be used to
demostrate the capabilities of the Next-Generation Clustered Heat Map (NG-CHM) system.
}
\details{
The included data is a small subset of data from The Cancer Genome Atlas (TCGA) project.
There are five data files from three groups of cancer samples:
\itemize{
\item 200 breast cancer (BRCA) samples.
\item 169 glioblastoma (GBM) samples.
\item 547 additional glioblastoma (GBM) samples.
}
The two Glioblastoma groups were characterized using different technologies (RNASeq and microarrays, respectively).

\strong{Note:} the NG-CHM system can work with data from any domain, not just biological data.

\strong{Note:} the included data has been been preprocessed, subsetted, and manipulated in multiple, undocumented ways. It
should only be used for evaluating and demonstrating the NG-CHM system and not for deriving any scientific conclusions.

The different data sets overlap with each other in several ways that are documented in the data sets concerned.  These
overlaps can be easily used to generate NG-CHMs that integrate multiple data sets in a variety of ways to further
demonstrate the capabilities of NG-CHMs.
}
\seealso{
TCGA.BRCA.Demo, TCGA.GBM.Demo, TCGA.GBM.EXPR
}
