% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.BRCA.Demo}
\alias{TCGA.BRCA.Demo}
\title{A subset of the breast cancer (BRCA) data from TCGA.}
\description{
This dataset is loaded automatically when the package is loaded. It
consists of two related parts:
\itemize{
\item A matrix of gene expression data.
\item A vector containing the TP53 mutation status of each sample in the matrix.
}
}
\seealso{
TCGA.BRCA.ExpressionData, TCGA.BRCA.TP53MutationData
}
