% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.BRCA.ExpressionData}
\alias{TCGA.BRCA.ExpressionData}
\title{A subset of the breast cancer (BRCA) expression data from TCGA}
\format{
A numeric data matrix with 3437 rows and 200 columns.
\itemize{
\item Row labels are gene symbols (e.g. TSPAN6).  The NG-CHM label type is \code{bio.gene.hugo}.
\item Column labels are TCGA barcodes up to the sample/vial field (16 characters total, e.g. TCGA-AO-A0JJ-01A). The NG-CHM label type is \code{bio.tcga.barcode.sample.vial}.
\item Data has been log-transformed (min 1, max 21.75322)
}
}
\description{
A subset of the breast cancer (BRCA) expression data from TCGA
}
\seealso{
TCGA.BRCA.Demo, TCGA.BRCA.TP53MutationData
}
