% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.BRCA.TP53MutationData}
\alias{TCGA.BRCA.TP53MutationData}
\title{TP53 mutation data for TCGA breast cancer (BRCA) samples}
\format{
A length 200 character vector.
\itemize{
\item Each element of the vector is either "WT" or "MUT".
\item Element names are TCGA barcodes up to the sample/vial field (16 characters total, e.g. TCGA-AO-A0JJ-01A)
}
}
\description{
TP53 mutation data for TCGA breast cancer (BRCA) samples
}
\seealso{
TCGA.BRCA.Demo, TCGA.BRCA.ExpressionData
}
