% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.GBM.Demo}
\alias{TCGA.GBM.Demo}
\title{A subset of the glioblastoma mutliforme (GBM) data from TCGA.}
\description{
This dataset is loaded by calling \code{data(TCGA.GBM.Demo)}.
}
\details{
The loaded data consists of two related parts:
\itemize{
\item A matrix of gene expression data.
\item A vector containing the TP53 mutation status of each sample in the matrix.
}
}
\seealso{
TCGA.GBM.ExpressionData, TCGA.GBM.TP53MutationData
}
