% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.GBM.TP53MutationData}
\alias{TCGA.GBM.TP53MutationData}
\title{TP53 mutation data for TCGA glioblastoma multiforme (GBM) samples}
\format{
A length 169 character vector.
\itemize{
\item Each element of the vector is either "WT" or "MUT".
\item Element names are TCGA barcodes up to the center field (28 characters total, e.g. TCGA-06-0178-01A-01R-1849-01).
}
}
\description{
Load using .
}
\seealso{
TCGA.GBM.Demo, TCGA.GBM.ExpressionData
}
